requirements.txt
Flask
pymongo
requests
selenium
python-dotenv